Added a confirmation alert before deleting a game

+++ src/components/games/FreeGamesList.js
@@ -75,7 +75,9 @@ const FreeGamesList = () => {
   }, [wasWonSuccess, deleteSuccess]);
 
   const handleDelete = (id) => {
-    dispatch(deleteGame(id));
+    if (window.confirm("Are you sure you want to delete this game?")) {
+      dispatch(deleteGame(id));
+    }
   };
 
   return (
